Keep Talking and Nobody Explodes Replica
This arduino code is part of a project to create a safe, physical
replica of the game of "Keep Talking and Nobody Explodes,"
(http://www.keeptalkinggame.com/) a cooperative video game where players
attempt to defuse a "bomb" in a set amount of time.

The "bomb" in the game has 14+ "modules" that must be "defused". Each
module requires a small microprocessor set up as a "slave" to a central
processor set up as a "master". The master branch will contail code for
the "master" processor, and secondary branches will contain code for the
individual modules.